Split `monument_cli` into a combined lib/bin
